Fix issues with racy SimpleClientResponseAdapter raciness as well as rollback issues in sysprocs

+++ src/frontend/org/voltdb/StoredProcedureInvocation.java
@@ -152,7 +152,7 @@ public class StoredProcedureInvocation implements FastSerializable, JSONString {
         return null;
     }
 
-    public void setClientHandle(int aHandle) {
+    public void setClientHandle(long aHandle) {
         clientHandle = aHandle;
     }
